create order and updating stock using observer and service repo pattern.

+++ database/migrations/2022_11_08_081025_create_ingredient_product_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->uuid('product_id');
             $table->uuid('ingredient_id');
-            $table->integer('ingredient_amount');
+            $table->double('ingredient_amount');
 
             $table->foreign('product_id')->references('id')->on('products')->cascadeOnDelete();
             $table->foreign('ingredient_id')->references('id')->on('ingredients')->cascadeOnDelete();
